Rename sample to app
Renaming sample package to app
